Fix Building Structure

+++ include/iconv/loop_unicode.h
@@ -44,7 +44,7 @@ static int unicode_transliterate (conv_t cd, ucs4_t wc,
         sub_outcount = cd->ofuncs.xxx_wctomb(cd,outptr,buf[i],outleft);
         if (sub_outcount <= RET_ILUNI)
           goto johab_hangul_failed;
-        if (!(sub_outcount <= outleft)) abort();
+        if (!(sub_outcount <= (int)(outleft))) abort();
         outptr += sub_outcount; outleft -= sub_outcount;
       }
       return outptr-backup_outptr;
@@ -88,7 +88,7 @@ static int unicode_transliterate (conv_t cd, ucs4_t wc,
             sub_outcount = cd->ofuncs.xxx_wctomb(cd,outptr,buf[i],outleft);
             if (sub_outcount <= RET_ILUNI)
               goto variant_failed;
-            if (!(sub_outcount <= outleft)) abort();
+            if (!(sub_outcount <= (int)outleft)) abort();
             outptr += sub_outcount; outleft -= sub_outcount;
           }
           return outptr-backup_outptr;
@@ -139,7 +139,7 @@ static int unicode_transliterate (conv_t cd, ucs4_t wc,
           sub_outcount = unicode_transliterate(cd,cp[i],outptr,outleft);
         if (sub_outcount <= RET_ILUNI)
           goto translit_failed;
-        if (!(sub_outcount <= outleft)) abort();
+        if (!(sub_outcount <= (int)outleft)) abort();
         outptr += sub_outcount; outleft -= sub_outcount;
       }
       return outptr-backup_outptr;
@@ -255,7 +255,7 @@ static void mb_to_uc_write_replacement (const unsigned int *buf, size_t buflen,
       if (cd->hooks.uc_hook)
         (*cd->hooks.uc_hook)(wc, cd->hooks.data);
       #endif
-      if (!(outcount <= outleft)) abort();
+      if (!(outcount <= (int)outleft)) abort();
       outptr += outcount; outleft -= outcount;
     outcount_zero: ;
     }
@@ -415,11 +415,11 @@ static size_t unicode_loop_convert (iconv_t icd,
       if (cd->hooks.uc_hook)
         (*cd->hooks.uc_hook)(wc, cd->hooks.data);
       #endif
-      if (!(outcount <= outleft)) abort();
+      if (!(outcount <= (int)outleft)) abort();
       outptr += outcount; outleft -= outcount;
     }
   outcount_zero:
-    if (!(incount <= inleft)) abort();
+    if (!(incount <= (int)inleft)) abort();
     inptr += incount; inleft -= incount;
   }
   *inbuf = (const char*) inptr;
@@ -500,7 +500,7 @@ static size_t unicode_loop_reset (iconv_t icd,
         if (cd->hooks.uc_hook)
           (*cd->hooks.uc_hook)(wc, cd->hooks.data);
         #endif
-        if (!(outcount <= outleft)) abort();
+        if (!(outcount <= (int)outleft)) abort();
         outptr += outcount;
         outleft -= outcount;
       outcount_zero:
@@ -516,7 +516,7 @@ static size_t unicode_loop_reset (iconv_t icd,
         errno = E2BIG;
         return -1;
       }
-      if (!(outcount <= outleft)) abort();
+      if (!(outcount <= (int)outleft)) abort();
       *outbuf = (char*) (outptr + outcount);
       *outbytesleft = outleft - outcount;
     }
