+++ include/iconv/utf7.h
@@ -159,7 +159,7 @@ active:
         else
           break;
       }
-      if (n < count+base64count+1)
+      if (n < (int)(count+base64count+1))
         goto none;
     }
     /* Here k = kmax > 0, hence base64count > 0. */
@@ -328,7 +328,7 @@ utf7_reset (conv_t conv, unsigned char *r, int n)
   if (state & 3) {
     /* deactivate base64 encoding */
     unsigned int count = ((state & 3) >= 2 ? 1 : 0) + 1;
-    if (n < count)
+    if (n < (int)count)
       return RET_TOOSMALL;
     if ((state & 3) >= 2) {
       unsigned int i = state & -4;
